requirements.txt
certifi==2017.4.17
chardet==3.0.4
geoip2==2.5.0
idna==2.5
maxminddb==1.3.0
msgpack-python==0.4.8
python-geoip==1.2
requests==2.18.1
urllib3==1.21.1
